+++ src/core/launcher/createStore.js
@@ -1,6 +1,9 @@
 import combineEpics from '@home/core/combineEpics';
 import { applyMiddleware, compose, createStore } from 'redux';
 import { createEpicMiddleware } from 'redux-observable';
+import { persistReducer } from 'redux-persist'
+import storage from 'redux-persist/lib/storage'
+import autoMergeLevel2 from 'redux-persist/lib/stateReconciler/autoMergeLevel2'
 
 export const stores = {};
 
@@ -11,7 +14,8 @@ export default function configureStore({
     enhancers = [],
     reducers,
 }) {
-    return function ({ name, http }) {
+    return function ({ name, http, persist }) {
+        // 处理epics
         let epicMiddleware;
         let mergedMiddlewares;
         if (epics) {
@@ -26,7 +30,24 @@ export default function configureStore({
         const mergedEnhancers = [middlewareEnhancer, ...enhancers];
         const composedEnhancers = compose(...mergedEnhancers);
 
-        const store = createStore(reducers, preloadedState, composedEnhancers);
+        // 处理persist
+        let reducersTrans = reducers;
+        let persistConfig;
+        if (persist) {
+            persistConfig = {
+                key: name,
+                storage: storage,
+                stateReconciler: autoMergeLevel2,
+                ...(persist === true ? {} : persist),
+            };
+            reducersTrans = persistReducer(persistConfig, reducers);
+        }
+
+        const store = createStore(
+            reducersTrans,
+            preloadedState,
+            composedEnhancers
+        );
 
         if (epics) {
             const originalEpics = combineEpics(http)(...epics);