+++ src/page/room/app/index.js
@@ -27,4 +27,5 @@ export default Launcher.createApp({
     },
     component: PageEntry,
     Http: Http,
+    persist: true,
 });